feature: move App and add routes
