--- src/App.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-
-      </header>
-    </div>
-  );
-}
-
-export default App;
