+++ src/app/App.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import './App.css';
+import {Route, Routes} from "react-router-dom";
+import {Main} from "../components/Main";
+
+
+function App() {
+
+    return (
+        <Routes>
+            <Route path="/" element={<Main/>}/>
+        </Routes>
+    );
+}
+
+export default App;
